+++ src/ListItems.css
@@ -5,11 +5,13 @@
     border: 2px solid salmon;
     border-radius: 5px;
     cursor: pointer;
+    transition: background-color 0.5s ease;
 }
 
 .list-item:hover {
     background-color: rgba(250, 128, 114, 0.4);
     color: white;
+    
 }
 
 .list-item:active {
